==> Backend (ML)/2.2 Model Training.ipynb <==
# Import the required modules
import keras
import pandas as pd
from keras.models import Sequential
from keras.layers import Dense, Dropout, Flatten
from keras.layers import Conv2D, MaxPooling2D
import keras.utils as image
from keras.callbacks import EarlyStopping
from keras.callbacks import ModelCheckpoint
from keras.preprocessing.image import ImageDataGenerator
import numpy as np 
from tqdm import tqdm
import cv2
from keras import models
from keras import layers
from keras import optimizers
from PIL import Image
from keras.applications import VGG16
import warnings
warnings.filterwarnings('ignore')

# Function to load the image, and convert into np arrays, to further train the model
def arrange_data(df):
    
    image_data = []
    img_paths = np.asarray(df.iloc[:, 0]) #First column is the image paths
    
    for i in tqdm(range(len(img_paths))):       
        img = image.load_img(img_paths[i],target_size=(200,150,3))
        img = image.img_to_array(img)
        img = img/255
        image_data.append(img)
        
    X = np.array(image_data)
    Y = np.array(df.iloc[:,1:29])
    
    # Display the size
    print("Shape of images:", X.shape)
    print("Shape of labels:", Y.shape)
    
    return X, Y

pd.set_option('display.max_colwidth', None)

train_df.iloc[:,1:29]

#Read and print and call the fucntion
print("Processing train..")
train_df = pd.read_csv("Data/Train.csv", delimiter=" ")
X_train, Y_train = arrange_data (train_df)

print("Processing valid..")
val_df = pd.read_csv("Data/Valid.csv", delimiter=" ")
X_val, Y_val = arrange_data (val_df)

#Number of classes
num_classes = 28  

#Model Declaration
model = Sequential()
model.add(Conv2D(filters=16, kernel_size=(5, 5), activation="relu", input_shape=(200,150,3)))
model.add(MaxPooling2D(pool_size=(2, 2)))
model.add(Dropout(0.25))
model.add(Conv2D(filters=32, kernel_size=(5, 5), activation='relu'))
model.add(MaxPooling2D(pool_size=(2, 2)))
model.add(Dropout(0.25))
model.add(Conv2D(filters=64, kernel_size=(5, 5), activation="relu"))
model.add(MaxPooling2D(pool_size=(2, 2)))
model.add(Dropout(0.25))
model.add(Conv2D(filters=64, kernel_size=(5, 5), activation='relu'))
model.add(MaxPooling2D(pool_size=(2, 2)))
model.add(Dropout(0.25))
model.add(Flatten())
model.add(Dense(128, activation='relu'))
model.add(Dropout(0.5))
model.add(Dense(64, activation='relu'))
model.add(Dropout(0.5))
model.add(Dense(num_classes, activation='sigmoid'))
model.summary()

#Model Training and Saving
model.compile(optimizer='adam', loss='binary_crossentropy', metrics=['accuracy'])
history = model.fit(X_train, Y_train, epochs=10, validation_data=(X_val, Y_val), batch_size=64)
model.save('Models/Model_6c.h5')

# Number of Genres
num_classes = 28

#Model Declaration
vgg_conv = VGG16(weights='imagenet', include_top=False, input_shape=(200,150, 3))
for layer in vgg_conv.layers[:-4]:
    layer.trainable = False
    
model = models.Sequential()
model.add(vgg_conv)
model.add(layers.Flatten())
model.add(layers.Dense(1024, activation='relu'))
model.add(layers.Dropout(0.5))
model.add(layers.Dense(num_classes, activation='sigmoid'))
model.summary()

#Model Training and Saving
model.compile(optimizer=optimizers.RMSprop(lr=1e-4), loss='binary_crossentropy', metrics=['accuracy'])
aug = ImageDataGenerator(rotation_range=20, zoom_range=0.15,width_shift_range=0.2, height_shift_range=0.2, 
                         shear_range=0.15,horizontal_flip=True, fill_mode="nearest")

EPOCHS=10
BS = 64
history = model.fit_generator(aug.flow(X_train, Y_train, batch_size=BS),validation_data=(X_val, Y_val), 
                    steps_per_epoch=len(X_train) // BS, epochs=EPOCHS)
model.save('Models/Model_4d.h5')
