==> Backend (ML)/3.3 Prediction & Analysis.ipynb <==
import pandas as pd
import numpy as np
import warnings
warnings.filterwarnings('ignore')
from sklearn.metrics import accuracy_score
from sklearn.feature_extraction.text import TfidfVectorizer
import pickle
from random import randint
import pandas as pd
import numpy as np
import warnings
warnings.filterwarnings('ignore')
from sklearn.model_selection import train_test_split
from sklearn.feature_extraction.text import TfidfVectorizer
from sklearn.svm import SVC
from sklearn.metrics import accuracy_score
from sklearn.naive_bayes import MultinomialNB
import pickle

test = pd.read_csv('Data/cleaned.csv')

x_train, x_test, y_train, y_test = train_test_split(test['Synopsis'].values, test['Genre'].values, test_size = 0.2, random_state = 0)

vectorizer = TfidfVectorizer()
f = vectorizer.fit_transform(x_train)
ft = vectorizer.transform(x_test)

svm_model = pickle.load(open("Models/model_svm.sav", 'rb'))
y_pred1 = svm_model.predict(ft)
print(accuracy_score(y_test, y_pred1))

index = np.random.randint(low=0, high=len(x_test))
print("Text: ", x_test[index])
print("Prediction: ", y_pred1[index])

mnb_model = pickle.load(open("Models/model_mnb.sav", 'rb'))
y_pred2 = mnb_model.predict(ft)
print(accuracy_score(y_test, y_pred2))

index = np.random.randint(low=0, high=len(x_test))
print("Text: ", x_test[index])
print("Prediction", y_pred2[index])
